pyproject.toml
[tool.poetry]
name = "obrazy1"
version = "0.1.0"
description = ""
authors = ["Maciej Lyskawinski <lyskawinski.maciej@gmail.com>"]

packages = [
    { include = "scripts" }
]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.1.3"
Pillow = "^9.1.0"
numpy = "^1.22.3"
tqdm = "^4.64.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
normalize = "scripts.main:image_normalize"
filter = "scripts.main:image_filter"
close = "scripts.main:image_close"
convex = "scripts.main:image_convex"
